Initial web wrapper
This is a functioning web wrapper application which can, in theory, be
picked up and run with by swapping out the URL that gets loaded for
whatever you want.

There are a few things that need to improve.  The BIFFUD logo is not
perfectly aligned between the loading screen and the web view.  Also,
the random color generator always returns the same color.

If that color had been randomly selected then I guess technically it
would still be "random" in that sense, but the reality is I picked the
first color of the list...